requirements.txt
# FastAPI microservice project dependencies

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0

# HTTP client for inter-service communication
httpx==0.25.2

# Template engine removed - using React frontend instead

# YAML processing
PyYAML==6.0.1

# LLM service dependencies
dashscope>=1.14.0
erniebot>=0.4.0
requests>=2.31.0
tencentcloud-sdk-python>=3.0.0
websocket-client>=1.6.0
volcengine-python-sdk[ark]>=1.0.0

# TTS service dependencies
websockets>=12.0
Pillow>=10.0.0

# Document processing dependencies
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pymupdf>=1.23.0
python-pptx>=0.6.21
python-docx>=0.8.11
aiofiles>=23.2.0
motor>=3.3.0
pymongo>=4.6.0

# Slide processing dependencies
pdf2image>=1.16.3
Pillow>=10.0.0

# ASR service dependencies
baidu-aip>=4.16.0
python-multipart>=0.0.6

# Database related (using MongoDB for knowledge and prompt management)
# sqlalchemy==2.0.23
# databases[postgresql]==0.8.0
# psycopg2-binary==2.9.9
bson>=0.5.10

# Caching (optional)
# redis==5.0.1

# Logging
python-json-logger==2.0.7

# Configuration management
python-dotenv==1.0.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Documentation tools
mkdocs==1.5.3
mkdocs-material==9.4.8


loguru